Add the coin flip function
A coin flip was added at the start so that the game won't always start
with the AI player having the center square already claimed. From this
point onwards, work must be focused on implementing a proper AI to the
moves of the computer moves and to clean/reestructure code whenever needed
to keep it tidy.