Rework of the player enter_move() function
In previous versions the code had checks for numbers outside the range
already, but any input other than a number would cause a ValueError and
stop the game at any point. Now upon entering a string or anything that
isn't an integer, the game prints a warning and asks for a new input.

This rework had the added benefit of separating our types of errors in a
way where simply changing the text of previous print() functions allowed
to have specific warnings for every case. Probably can be achieved by
better use of the try and except blocks in the future, but this made for
an easy fix.