Add first check to minmax algorithm
The first chunk of our algorithm simply scans every possible next move
to see if it can win the game with it, and it it has the chance to do so
it picks that winning square. Had a bit of a trouble copying the list
containing the current state of the board to use internally: as it was a
list of lists, even though we copied it correctly (using the "[:]" trick)
the sublists would remain linked and changes to the internal_board variable
would leak into our current_board. Importing the copy module and using
deepcopy instead fixed all issues and the machine now can pick the correct
move as long as it has a "checkmate" in the next turn