Add detection of player win in next move
We have added that detects if the human player can win in the next turn
then tries to prevent it by blocking the first detected win move. Testing
shows the game now has an appropiate feel to AI behavior even without
use of the minmax algorithm we were researching. Game behavior could
be kept in its current state and focus instead on QOL and visual updates.

Two ideas on improvements at this point are to add the classic main title
to the game and a PLAY or QUIT select screen and an animation for the
starting coin toss. These should be prioritized over AI at this point

Once the game and the code is looking better, then and only then should
we implement the fill minmax algorithm as a final research on its
functionality in Python.