Add Main Menu file, add pixelated font, start pygame committed
I've added a separate file for the main menu, installed pygame and
started to use it to render the main menu of our game. The skeleton code
so far only draws a "MAIN MENU" title on an empty black screen. We should
expand on it with buttons:
    * A PLAY button that starts the game
    * An OPTION button that opens a new screen to resize the window
    * A QUIT button that closes the screen
We should also create a background image. The font can be kept or changed
at a later time. Reading of pygame documentation should be prioritized